LayoutTests/platform/ios-simulator/editing/pasteboard/paste-list-004-expected.txt

CONSOLE MESSAGE: line 54: Unexpected list item: 1,three,three 
Pasting a list item into the middle of another list item should split the target list item into two with the pasted content in between.

one two
three 
one two
four
monsters walking 
one two
cross the floor
FAIL
